requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LLM & RAG
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-chroma>=0.2.0
langchain-openai>=0.3.0
chromadb>=0.5.0
ollama>=0.4.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.12.0

# Search
rank-bm25>=0.2.2

# Embeddings
sentence-transformers>=3.0.0

# Utilities
loguru>=0.7.0
httpx>=0.27.0
tenacity>=8.2.0
tiktoken>=0.7.0
